Changes HAL to return -1 to 1 for joysticks
First part of the HAL changes. Returns -1 to 1 for joysticks instead of
-128 to 127. Implemented by replacing the old structure with a new
structure that uses floats instead of shorts

+++ wpilibj/src/athena/java/edu/wpi/first/wpilibj/DriverStation.java
@@ -31,11 +31,11 @@ public class DriverStation implements RobotState.Interface {
   }
 
   private class HALJoystickAxes {
-    public short[] m_axes;
+    public float[] m_axes;
     public byte m_count;
 
     public HALJoystickAxes(int count) {
-      m_axes = new short[count];
+      m_axes = new float[count];
     }
   }
